Split up the curves one per module
