+++ src/lib/math/pcurves/info.txt
@@ -20,4 +20,5 @@ pcurves_id.h
 pcurves_impl.h
 pcurves_util.h
 pcurves_wrap.h
+pcurves_instance.h
 </header:internal>
