+++ src/lib/math/pcurves/pcurves_brainpool256r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_BRAINPOOL256R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve brainpool256r1"
+brief -> "brainpool256r1"
+type -> "Internal"
+</module_info>
