+++ src/lib/math/pcurves/pcurves_brainpool384r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_BRAINPOOL384R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve brainpool384r1"
+brief -> "brainpool384r1"
+type -> "Internal"
+</module_info>
