+++ src/lib/math/pcurves/pcurves_brainpool512r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_BRAINPOOL512R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve brainpool512r1"
+brief -> "brainpool512r1"
+type -> "Internal"
+</module_info>
