+++ src/lib/math/pcurves/pcurves_impl.h
@@ -64,70 +64,6 @@ class MontgomeryRep final {
       }
 };
 
-template <typename Params>
-class P521Rep final {
-   private:
-      template <WordType W, size_t N>
-      static consteval bool is_p521(const std::array<W, N>& p) {
-         if constexpr(N != (521 + WordInfo<W>::bits - 1) / WordInfo<W>::bits) {
-            return false;
-         }
-
-         if(p[N - 1] != 0x1FF) {
-            return false;
-         }
-         for(size_t i = 0; i != N - 1; ++i) {
-            if(~p[i] != 0) {
-               return false;
-            }
-         }
-
-         return true;
-      }
-
-   public:
-      static constexpr auto P = Params::P;
-      static constexpr size_t N = Params::N;
-      typedef typename Params::W W;
-
-      static_assert(is_p521(P));
-
-      constexpr static std::array<W, N> one() {
-         std::array<W, N> one = {};
-         one[0] = 1;
-         return one;
-      }
-
-      constexpr static std::array<W, N> redc(const std::array<W, 2 * N>& z) {
-         constexpr W TOP_MASK = static_cast<W>(0x1FF);
-
-         std::array<W, N> hi = {};
-         copy_mem(hi, std::span{z}.template subspan<N - 1, N>());
-         shift_right<9>(hi);
-
-         std::array<W, N> lo = {};
-         copy_mem(lo, std::span{z}.template first<N>());
-         lo[N - 1] &= TOP_MASK;
-
-         // s = hi + lo
-         std::array<W, N> s = {};
-         // Will never carry out
-         W carry = bigint_add<W, N>(s, lo, hi);
-
-         // But might be greater than modulus:
-         std::array<W, N> r = {};
-         bigint_monty_maybe_sub<N>(r.data(), carry, s.data(), P.data());
-
-         return r;
-      }
-
-      constexpr static std::array<W, N> to_rep(const std::array<W, N>& x) { return x; }
-
-      constexpr static std::array<W, N> wide_to_rep(const std::array<W, 2 * N>& x) { return redc(x); }
-
-      constexpr static std::array<W, N> from_rep(const std::array<W, N>& z) { return z; }
-};
-
 template <typename Rep>
 class IntMod final {
    private:
