+++ src/lib/math/pcurves/pcurves_secp256k1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_SECP256K1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve secp256k1"
+brief -> "secp256k1"
+type -> "Internal"
+</module_info>
