+++ src/lib/math/pcurves/pcurves_secp256r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_SECP256R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve secp256r1"
+brief -> "secp256r1"
+type -> "Internal"
+</module_info>
