+++ src/lib/math/pcurves/pcurves_secp256r1/pcurves_secp256r1.cpp
@@ -1,0 +1,41 @@
+/*
+* (C) 2024 Jack Lloyd
+*
+* Botan is released under the Simplified BSD License (see license.txt)
+*/
+
+#include <botan/internal/pcurves_instance.h>
+
+#include <botan/internal/pcurves_wrap.h>
+
+namespace Botan::PCurve {
+
+namespace {
+
+// clang-format off
+namespace secp256r1 {
+
+class Params final : public EllipticCurveParameters<
+   "FFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFF",
+   "FFFFFFFF00000001000000000000000000000000FFFFFFFFFFFFFFFFFFFFFFFC",
+   "5AC635D8AA3A93E7B3EBBD55769886BC651D06B0CC53B0F63BCE3C3E27D2604B",
+   "FFFFFFFF00000000FFFFFFFFFFFFFFFFBCE6FAADA7179E84F3B9CAC2FC632551",
+   "6B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296",
+   "4FE342E2FE1A7F9B8EE7EB4A7C0F9E162BCE33576B315ECECBB6406837BF51F5",
+   -10> {
+};
+
+class Curve final : public EllipticCurve<Params> {};
+
+}
+
+// clang-format on
+
+}  // namespace
+
+std::shared_ptr<const PrimeOrderCurve> PCurveInstance::secp256r1() {
+   static auto g_secp256r1 = std::make_shared<const PrimeOrderCurveImpl<secp256r1::Curve>>();
+   return g_secp256r1;
+}
+
+}  // namespace Botan::PCurve
