+++ src/lib/math/pcurves/pcurves_secp384r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_SECP384R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve secp384r1"
+brief -> "secp384r1"
+type -> "Internal"
+</module_info>
