+++ src/lib/math/pcurves/pcurves_secp521r1/info.txt
@@ -1,0 +1,9 @@
+<defines>
+PCURVES_SECP521R1 -> 20240608
+</defines>
+
+<module_info>
+name -> "PCurve secp521r1"
+brief -> "secp521r1"
+type -> "Internal"
+</module_info>
